Can now create/edit users successfully from GUI
Previously, the password hash was always being
recomputed when updating, even if user didn't specify
password.

On save, the validate() method was not being called
validate() and save() actually check the constraints

The hasErrors() method only check the data bindings.
Therefore, if the user had errors, the program
still tried to save, and then still tried to render
the show page.

Closes #58