Fixed bug with creating new Lessons
When user would change the name of the new lesson,
the entire <li> would be replaced.  This would
zap the lessonId_NEW_KID_ON_THE_BLOCK form field.

A new <div> tag was given the id of newLessonNameInSequence,
so that the entire <li> wouldn't be blasted when the user
typed in the name of the new Lesson.

See also createLesson.js